pyproject.toml
[tool.poetry]
name = "stellate-strawberry"
version = "1.1.1"
description = "Integrate Stellate with your Strawberry GraphQL API"
authors = ["Stellate <eng@stellate.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
strawberry-graphql = "*"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
